Move test helper plugins to testing package
